backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
pydantic[email]>=2.3.0
email-validator>=2.0.0
sqlalchemy>=2.0.0
python-jose>=3.3.0
passlib[argon2]>=1.7.4
argon2-cffi>=21.3.0
python-multipart>=0.0.9

# ML stack used by ai/serve_agri.py
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
xgboost>=1.7.0
